Extracted RestRequestBatchKey class from ParSeqRestClient.
